requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
pydantic==2.6.0
python-dateutil==2.8.2
google-generativeai==0.3.1
pytest==7.4.3